Split out rate limiter.
